+++ RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.1.0] .
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/circlemind-sdk/0.1.0 - .
+
+## 2024-12-16 16:50:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.455.6 (2.479.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.0-post0] .
+### Releases
+- [PyPI v0.1.0-post0] https://pypi.org/project/circlemind-sdk/0.1.0-post0 - .